# Removing defer in hot code - incCollected and incEmitted