@yosiat 2016-04-02
I closed #402 because I accidently created the branch based on fork-by-measurement, this is now on based on master.