@yosiat 2016-04-04
@nathanielc by the way, I found out a lot of methods (which aren't hot) with one line, like this:

```
lock()
defer unlock()
return ..
```

I can change those, but this can too much aggresive