@nathanielc 2016-04-04
Yeah, let's leave those alone since they are not hot.
On Apr 4, 2016 12:14 PM, "Yosi Attias" <notifications@github.com> wrote:

> @nathanielc <https://github.com/nathanielc> by the way, I found out a lot
> of methods (which aren't hot) with one line, like this:
>
> lock()
> defer unlock()
> return ..
>
> I can change those, but this can too much aggresive
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub
> <https://github.com/influxdata/kapacitor/pull/403#issuecomment-205428173>
>